init: Extract dump part from FindPartitionsToUmount
Currently, for the FindPartitionsToUmount() function is not doing what
the function name and comments describe. If the dump argument is true,
it won't find partitions to umount. Instead, it just log out all mount
points in /proc/mounts.

This CL creates a new function DumpPartitions() to log out all mount
points in /proc/mounts, and delete dump argument in
FindPartitionsToUmount() function.

Bug: 409196828
Test: Verify umount in reboot works well by checking pstore logs
Test: Verify the mount points are dumped when umount fails
Change-Id: I3644af8b772dfcb42dc4b6724811778ea124892f