fastboot: Delay resparsing until flash_buf() is called.
load_buf() tries to resparse a file if the max-download-size is smaller
than the number of bytes to transfer. But it's buggy, because there are
multiple places where we rewrite the buffer, but do not check if we need
to resparse. It's also confusing because we have to track the underlying
file type as well as the actual buffer type.

Instead, move the resparsing logic directly into flash_buf(), which is
the only place the resparsing was ever used.

Note: "fastboot stage" no longer attempts to resparse, because it calls
Download() directly and never actually used the resparse results.

Bug: 415214175
Test: fastboot flashall
Change-Id: I1c2af7630e86301b6388bc07a58634051e53a756