capture pstore logs in init
When a device OTA's and fails to boot, after several attempts, it will
rollback. Once this happens, /data is wipe and we no longer have any
forensic information about the boot failure. pstore is a linux
persistent storage that captures some of these logs. To better diagnose
rollback issues, we should preserve the logs from pstore once we detect
a rollback in the boot process (as early as possible), hold it in RAM
and then flush it out to /data/misc/update_engine_logs once data is
mountedto be captured by bugreport.

Bug: 394172800
Test: th
Change-Id: I7a2b79d7bb53a5ed939c4802b95442f6a9fd5dc9