Use Protobuf without TLS for linker dependencies.
The Bionic linker does not support ELF TLS when it's loading
itself, so it cannot have any library that uses TLS in its
transitive dependencies. Use a variant of Protobuf that disables
TLS as a workaround.

The long-term fix would be to add TLS-in-linker support. See
b/431861292 for current status.

Bug: 329747255
Test: booted emulator and Pixel devices, presubmit
Flag: EXEMPT third-party library upgrade
Change-Id: I833f50a45e57be2e0b511b936d84fe7fb25db1f2