libmodprobe: Do not log error of finit_module(2) on EEXIST
Current code logged an error message when finit_module(2) failed
regardless of the reason of the failure. This caused a regression in
terms of the logcat size as modprobe did not log error on EEXIST before
ag/34337743. This commit adds a simple if clause to align with the
original implementation.

Test: build passes, and log is not emitted on EEXIST
Bug: 433984587
Change-Id: Ibc2b42e6aeda299ac92e0aa0576d49a77cb6ba00